move helper out of escape-hatch-strategy.tsx
